Fix end time, add logging, and add support for Queens
There was an issue where the end time for the last day of contests
was ending at 11:59 am, instead of 11:59 pm. I fixed that and now
both Wordle and Queens scores will be accurate.

I added more logging so that it will be easier to spot issues in
the future, and to find the 11:59 am issue.

I generified stuff to add support for the LinkedIn Queens game,
and this should make it relatively easy to support other games in
the future as well.